pyproject.toml
[project]
name = "azure-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-core>=1.37.0",
    "azure-identity>=1.25.1",
    "azure-search>=1.0.0b2",
    "azure-search-documents>=11.6.0",
    "bokeh>=3.8.1",
    "fastapi>=0.127.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langgraph>=1.0.5",
    "ollama>=0.6.1",
    "openai>=2.14.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "setuptools>=80.9.0",
    "uvicorn>=0.40.0",
]
